Partiel CloudSecurityIot/Device-Management-v2/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
python-dotenv==1.0.1
paho-mqtt==1.6.1
python-jose==3.3.0
psutil==6.1.0
prometheus-fastapi-instrumentator==7.0.0
requests==2.32.3

